Test connection fixed
I'm loading the app for the first time. I want to setup supabase connection.
I enter the endpoint and the anon key, then I click on "Test COnnection". I get this: "Failed to set up Supabase tables. You may need to run these SQL commands manually in the Supabase dashboard SQL editor:".

I just want to test the connection, and that my app can connect to supabase. I want to change this to a status message that says something like: "Connection to supabase is successfull, but you need to initialize your databases by clicking on Initialize Schema".

Test Connection is just to make sure the endpoint and credentials are correct